Move query related logic to query package
Move all query related logic into its own package.
