+++ src/main/java/org/opensearch/knn/index/memory/NativeMemoryAllocation.java
@@ -12,6 +12,7 @@
 package org.opensearch.knn.index.memory;
 
 import org.apache.lucene.index.LeafReaderContext;
+import org.opensearch.knn.index.query.KNNWeight;
 import org.opensearch.knn.jni.JNIService;
 import org.opensearch.knn.index.util.KNNEngine;
 import org.opensearch.watcher.FileWatcher;
@@ -158,7 +159,7 @@ public interface NativeMemoryAllocation {
 
         /**
          * The read lock will be obtained in the
-         * {@link org.opensearch.knn.index.KNNWeight#scorer(LeafReaderContext context) scorer} when a native index needs
+         * {@link KNNWeight#scorer(LeafReaderContext context) scorer} when a native index needs
          * to be queried.
          */
         @Override
