+++ src/main/java/org/opensearch/knn/jni/FaissService.java
@@ -12,7 +12,7 @@
 package org.opensearch.knn.jni;
 
 import org.opensearch.knn.common.KNNConstants;
-import org.opensearch.knn.index.KNNQueryResult;
+import org.opensearch.knn.index.query.KNNQueryResult;
 import org.opensearch.knn.index.util.KNNEngine;
 
 import java.security.AccessController;
