+++ src/main/java/org/opensearch/knn/plugin/script/KNNScoringSpace.java
@@ -6,7 +6,7 @@
 package org.opensearch.knn.plugin.script;
 
 import org.opensearch.knn.index.mapper.KNNVectorFieldMapper;
-import org.opensearch.knn.index.KNNWeight;
+import org.opensearch.knn.index.query.KNNWeight;
 import org.apache.lucene.index.LeafReaderContext;
 import org.opensearch.index.mapper.MappedFieldType;
 import org.opensearch.script.ScoreScript;
