+++ src/test/java/org/opensearch/knn/index/KNNESSettingsTestIT.java
@@ -6,6 +6,7 @@
 package org.opensearch.knn.index;
 
 import org.opensearch.knn.KNNRestTestCase;
+import org.opensearch.knn.index.query.KNNQueryBuilder;
 import org.opensearch.knn.plugin.stats.StatNames;
 import org.apache.http.util.EntityUtils;
 import org.opensearch.client.Response;
