+++ src/test/java/org/opensearch/knn/index/codec/KNNCodecTestCase.java
@@ -13,10 +13,10 @@ import org.opensearch.knn.KNNTestCase;
 import org.opensearch.knn.common.KNNConstants;
 import org.opensearch.knn.index.codec.KNN920Codec.KNN920Codec;
 import org.opensearch.knn.jni.JNIService;
-import org.opensearch.knn.index.KNNQuery;
+import org.opensearch.knn.index.query.KNNQuery;
 import org.opensearch.knn.index.KNNSettings;
 import org.opensearch.knn.index.mapper.KNNVectorFieldMapper;
-import org.opensearch.knn.index.KNNWeight;
+import org.opensearch.knn.index.query.KNNWeight;
 import org.opensearch.knn.index.SpaceType;
 import org.opensearch.knn.index.VectorField;
 import org.apache.lucene.codecs.Codec;
