+++ src/test/java/org/opensearch/knn/index/memory/NativeMemoryLoadStrategyTests.java
@@ -17,7 +17,7 @@ import org.opensearch.action.search.SearchResponse;
 import org.opensearch.knn.KNNTestCase;
 import org.opensearch.knn.common.KNNConstants;
 import org.opensearch.knn.jni.JNIService;
-import org.opensearch.knn.index.KNNQueryResult;
+import org.opensearch.knn.index.query.KNNQueryResult;
 import org.opensearch.knn.index.SpaceType;
 import org.opensearch.knn.index.util.KNNEngine;
 import org.opensearch.knn.training.TrainingDataConsumer;
